fix:ファイル名の変更とインターフェースの切り分け
