+++ repository/game.go
@@ -1,0 +1,11 @@
+package repository
+
+import (
+	"winter_pj/model"
+)
+
+type GameRepository interface {
+	GetCompareImage(searchName string) (*model.ImageList, error)
+	InsertCompareImage(insertCompareImage model.InsertCompareImage) error
+	InsertFaceImage(insertFaceImage model.InsertFaceImage) error
+}
