+++ repository/game_sql.go
@@ -3,9 +3,3 @@ package repository
 import (
 	"winter_pj/model"
 )
-
-type GameRepository interface {
-	GetCompareImage(searchName string) (*model.ImageList, error)
-	InsertCompareImage(insertCompareImage model.InsertCompareImage) error
-	InsertFaceImage(insertFaceImage model.InsertFaceImage) error
-}
